2011-02-07  Ryosuke Niwa  <rniwa@webkit.org>
        Unreviewed Leopard build fix.

        * WebCore.xcodeproj/project.pbxproj: let XCode have developmentRegion.
        * platform/mac/DragImageMac.mm:
        (WebCore::webkit_CGCeiling): Cast the return value of ceil to CGFloat.


Canonical link: https://commits.webkit.org/67947@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@77888 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/mac/DragImageMac.mm
@@ -169,7 +169,7 @@ static inline CGFloat webkit_CGCeiling(CGFloat value)
 {
     if (sizeof(value) == sizeof(float))
         return ceilf(value);
-    return ceil(value);
+    return static_cast<CGFloat>(ceil(value));
 }
     
 static void drawAtPoint(NSString *string, NSPoint point, NSFont *font, NSColor *textColor)
